requirements.txt
selenium
pytest
python-dotenv
faker
allure-pytest